Move grpc error handling to separate file
